requirements.txt
selenium>=4.8.0
webdriver-manager>=4.0.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0